// material-vender-ui/src/app/models/vender.model.ts
import { PurchaseOrder } from "./purchase-order.model";

export interface Vendor {
    id: number;
    code: string;
    name: string;
    addressLine1?: string;
    addressLine2?: string;
    contactEmail?: string;
    contactNo?: string;
    validTillDate?: Date;
    isActive: boolean;
    purchaseOrders?: PurchaseOrder[];
  }
  